create TeamCreateOutputHeader component
